Rename file extension for typescript file
