+++ packages/react-meteor-data/withTracker.jsx
@@ -1,5 +1,5 @@
 import React, { forwardRef, memo } from 'react';
-import useTracker from './useTracker.js';
+import useTracker from './useTracker';
 
 export default function withTracker(options) {
   return Component => {
